requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
prophet>=1.1.0
matplotlib>=3.7.0
fastapi>=0.100.0
uvicorn>=0.23.0
apache-airflow>=2.7.0
shap>=0.42.0
seaborn>=0.12.0
xgboost>=2.0.0
plotly>=5.15.0
torch>=2.0.0
lime>=0.2.0
duckdb>=0.9.0
mlflow>=2.8.0
streamlit>=1.28.0
fpdf2>=2.7.0
pytest>=7.4.0
flake8>=6.1.0
